// src/types/doacaoAlimentoTypes.ts
export type DoacaoAlimentoCreateType = {
    id_solicitante: string;
    id_alimento: string;
    peso: string;
    id_condutor: string;
    id_estoque_alimento: string;
};

export type DoacaoAlimentoUpdateType = {
    id_solicitante: string;
    id_alimento: string;
    peso: string;
    id_condutor: string;
    id_estoque_alimento: string;
};